Fix a fairly significant bug.
This was basically a typo that was causing it to think that events
violated the "one event during Dojo hours" rule when they didn't.
The unit tests caught it. The reason they didn't do so earlier
was because the times just happened to work out earlier such that
it didn't manifest itself.